Add template HTML, basic CSS
Made the template HTML file with a simple div that will represent the
calculator later. Connected the files script.js and styles.css with the
index.html file, and made sure everything works for starters. There is
a basic beige background and a coloured div element that will become the
calculator.